refactor(daemon): extract orchestrator tool execution
